backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
sentence-transformers==3.1.1
numpy==2.3.5
pinecone-client==4.1.2
